+++ exporter/opentelemetry-exporter-otlp-proto-grpc/src/opentelemetry/exporter/otlp/proto/grpc/_log_exporter/__init__.py
@@ -28,7 +28,7 @@ from opentelemetry.proto.collector.logs.v1.logs_service_pb2 import (
 from opentelemetry.proto.collector.logs.v1.logs_service_pb2_grpc import (
     LogsServiceStub,
 )
-from opentelemetry.sdk._logs import LogData
+from opentelemetry.sdk._logs import ReadableLogRecord
 from opentelemetry.sdk._logs.export import LogExporter, LogExportResult
 from opentelemetry.sdk.environment_variables import (
     _OTEL_PYTHON_EXPORTER_OTLP_GRPC_LOGS_CREDENTIAL_PROVIDER,
@@ -46,7 +46,7 @@ from opentelemetry.sdk.environment_variables import (
 class OTLPLogExporter(
     LogExporter,
     OTLPExporterMixin[
-        Sequence[LogData],
+        Sequence[ReadableLogRecord],
         ExportLogsServiceRequest,
         LogExportResult,
         LogsServiceStub,
@@ -105,13 +105,13 @@ class OTLPLogExporter(
         )
 
     def _translate_data(
-        self, data: Sequence[LogData]
+        self, data: Sequence[ReadableLogRecord]
     ) -> ExportLogsServiceRequest:
         return encode_logs(data)
 
     def export(  # type: ignore [reportIncompatibleMethodOverride]
         self,
-        batch: Sequence[LogData],
+        batch: Sequence[ReadableLogRecord],
     ) -> Literal[LogExportResult.SUCCESS, LogExportResult.FAILURE]:
         return OTLPExporterMixin._export(self, batch)