+++ opentelemetry-sdk/src/opentelemetry/sdk/_events/__init__.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import logging
-import warnings
 from time import time_ns
 from typing import Optional
 
@@ -21,13 +20,13 @@ from opentelemetry import trace
 from opentelemetry._events import Event
 from opentelemetry._events import EventLogger as APIEventLogger
 from opentelemetry._events import EventLoggerProvider as APIEventLoggerProvider
-from opentelemetry._logs import NoOpLogger, SeverityNumber, get_logger_provider
-from opentelemetry.sdk._logs import (
-    LogDeprecatedInitWarning,
-    Logger,
-    LoggerProvider,
+from opentelemetry._logs import (
     LogRecord,
+    NoOpLogger,
+    SeverityNumber,
+    get_logger_provider,
 )
+from opentelemetry.sdk._logs import Logger, LoggerProvider
 from opentelemetry.util.types import _ExtendedAttributes
 
 _logger = logging.getLogger(__name__)
@@ -58,22 +57,17 @@ class EventLogger(APIEventLogger):
             return
         span_context = trace.get_current_span().get_span_context()
 
-        # silence deprecation warnings from internal users
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", category=LogDeprecatedInitWarning)
-
-            log_record = LogRecord(
-                timestamp=event.timestamp or time_ns(),
-                observed_timestamp=None,
-                trace_id=event.trace_id or span_context.trace_id,
-                span_id=event.span_id or span_context.span_id,
-                trace_flags=event.trace_flags or span_context.trace_flags,
-                severity_text=None,
-                severity_number=event.severity_number or SeverityNumber.INFO,
-                body=event.body,
-                resource=getattr(self._logger, "resource", None),
-                attributes=event.attributes,
-            )
+        log_record = LogRecord(
+            timestamp=event.timestamp or time_ns(),
+            observed_timestamp=None,
+            trace_id=event.trace_id or span_context.trace_id,
+            span_id=event.span_id or span_context.span_id,
+            trace_flags=event.trace_flags or span_context.trace_flags,
+            severity_text=None,
+            severity_number=event.severity_number or SeverityNumber.INFO,
+            body=event.body,
+            attributes=event.attributes,
+        )
         self._logger.emit(log_record)