+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- docs: Added sqlcommenter example
+  ([#4734](https://github.com/open-telemetry/opentelemetry-python/pull/4734))
+
 ## Version 1.38.0/0.59b0 (2025-10-16)
 
 - Add `rstcheck` to pre-commit to stop introducing invalid RST
@@ -41,8 +44,6 @@ inject a `requests.Session` or `grpc.ChannelCredentials` object into OTLP export
   ([#4735](https://github.com/open-telemetry/opentelemetry-python/pull/4735))
 - opentelemetry-sdk: fix calling Logger.emit with an API LogRecord instance
   ([#4741](https://github.com/open-telemetry/opentelemetry-python/pull/4741))
-- docs: Added sqlcommenter example
-  ([#4734](https://github.com/open-telemetry/opentelemetry-python/pull/4734))
 
 ## Version 1.36.0/0.57b0 (2025-07-29)