+++ CHANGELOG.md
@@ -12,12 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- docs: Added sqlcommenter example
+  ([#4734](https://github.com/open-telemetry/opentelemetry-python/pull/4734))
+
+## Version 1.38.0/0.59b0 (2025-10-16)
+
 - Add `rstcheck` to pre-commit to stop introducing invalid RST
   ([#4755](https://github.com/open-telemetry/opentelemetry-python/pull/4755))
 - logs: extend Logger.emit to accept separated keyword arguments
   ([#4737](https://github.com/open-telemetry/opentelemetry-python/pull/4737))
 - otlp exporters (trace): include W3C TraceFlags (bits 0–7) in OTLP `Span.flags` alongside parent isRemote bits (8–9)
   ([#4761](https://github.com/open-telemetry/opentelemetry-python/pull/4761))
+- logs: add warnings for classes that would be deprecated and renamed in 1.39.0
+  ([#4771](https://github.com/open-telemetry/opentelemetry-python/pull/4771))
 
 ## Version 1.37.0/0.58b0 (2025-09-11)